fix(mob): keep the wealth ledger in float32 under a half-precision model
``to(dtype)`` converted every floating buffer along with the parameters,
so under bf16 the ledgers were bf16 too. A wealth of 83.5 has a
resolution of 0.5 there while a step's transfer is of order 1e-2:
measured on Qwen3-1.7B over 120 steps the wealth vector never moved and
its standard deviation read exactly 0.0, which is the wrong reason for
an economy to look flat. The ledgers are recast to float32 after every
move; the auction reads wealth into the bid's dtype itself.

Refs: #15

+++ tame/mob/auction.py
@@ -187,7 +187,9 @@ class VCGAuctioneer(nn.Module):
         wealth: torch.Tensor,
     ) -> "AuctionOutcome":
         wealth_snapshot = wealth.detach().clone()
-        bids = confidences * wealth_snapshot.unsqueeze(0).unsqueeze(0)
+        # The ledger is float32 whatever the model runs in; the bid takes the
+        # report's dtype so the routing weights match the expert outputs they scale.
+        bids = confidences * wealth_snapshot.to(confidences.dtype).unsqueeze(0).unsqueeze(0)
         top_bids, selected_experts = torch.topk(bids, self.top_k, dim=-1)
         payments = self._compute_vcg_payments(bids, selected_experts, wealth_snapshot)
         rebates = self._compute_rebates(bids, wealth_snapshot)
